+++ src/lib.rs
@@ -52,6 +52,7 @@
 //! | [`cmul`][cmul] | `let result: Result<FixedPoint, ArithmeticError> = a.cmul(b)` | Checked multiplication. Returns `Err` on overflow. This is multiplication without rounding, hence it's available only when at least one operand is integer. |
 //! | [`rmul`][rmul] | `let result: Result<FixedPoint, ArithmeticError> = a.rmul(b, RoundMode::Ceil)` | Checked rounding multiplication. Returns `Err` on overflow. Because of provided [`RoundMode`][RoundMode] it's possible across the [`FixedPoint`][FixedPoint] values. |
 //! | [`rdiv`][rdiv] | `let result: Result<FixedPoint, ArithmeticError> = a.rdiv(b, RoundMode::Floor)` | Checked [rounding][RoundMode] division. Returns `Err` on overflow. |
+//! | [`rsqrt`][rsqrt] | `let result: Result<FixedPoint, ArithmeticError> = a.rsqrt(RoundMode::Floor)` | Checked [rounding][RoundMode] square root. Returns `Err` for negative argument. |
 //! | [`cneg`][cneg] | `let result: Result<FixedPoint, ArithmeticError> = a.cneg()` | Checked negation. Returns `Err` on overflow (you can't negate [`MIN` value][MIN]). |
 //! | [`integral`][integral] | `let y: {integer} = x.integral(RoundMode::Floor)` | Takes [rounded][RoundMode] integral part of the number. |
 //! | [`saturating_add`][saturating_add] | `let z: FixedPoint = x.saturating_add(y)` | Saturating addition |
@@ -111,6 +112,7 @@
 //! [parity_scale_codec]: https://docs.rs/parity-scale-codec
 //! [rdiv]: ./ops/trait.RoundingDiv.html#tymethod.rdiv
 //! [rmul]: ./ops/trait.RoundingMul.html#tymethod.rmul
+//! [rsqrt]: ./ops/trait.RoundingSqrt.html#tymethod.rsqrt
 //! [RoundMode]: ./ops/enum.RoundMode.html
 //! [saturating_add]: ./ops/trait.CheckedAdd.html#tymethod.saturating_add
 //! [saturating_mul]: ./ops/trait.CheckedMul.html#tymethod.saturating_mul
@@ -398,20 +400,16 @@ macro_rules! impl_fixed_point {
             type Error = ArithmeticError;
 
             #[inline]
-            fn rsqrt(self, mode: RoundMode) -> Result<Self> {
+            fn rsqrt(self, mode: RoundMode) -> Result<Self, Self::Error> {
+                if self.inner.is_negative() {
+                    return Err(ArithmeticError::DomainViolation);
+                }
                 // At first we have `S_inner = S * COEF`.
                 // We'd like to gain `sqrt(S) * COEF`:
                 // `sqrt(S) * COEF = sqrt(S * COEF^2) = sqrt(S_inner * COEF)`
                 let inner_squared = $promotion::from(self.inner) * Self::COEF_PROMOTED;
-                let inner_lo = inner_squared.sqrt()?;
-                let inner = match mode {
-                    RoundMode::Floor => inner_lo,
-                    RoundMode::Ceil => if inner_lo * inner_lo == inner_squared { inner_lo } else {
-                        inner_lo + $promotion::ONE
-                    },
-                }
-                    .try_into()
-                    .map_err(|_| ArithmeticError::Overflow)?;
+                let inner = inner_squared.rsqrt(mode)?;
+                let inner = inner.try_into().unwrap(); // `sqrt` can't take more bits than `self` already does
                 Ok(Self::from_bits(inner))
             }
         }
@@ -444,8 +442,8 @@ macro_rules! impl_fixed_point {
                 } else {
                     let min = a.inner.min(b.inner);
                     let max = a.inner.max(b.inner);
-                    let half_diff = Self::from_bits(max - min).rdiv(2, mode).unwrap();
-                    Self::from_bits(min + half_diff.inner)
+                    let half_diff = (max - min).rdiv(2, mode).unwrap();
+                    Self::from_bits(min + half_diff)
                 }
             }