+++ src/ops/mod.rs
@@ -1,7 +1,8 @@
-use core::convert::TryFrom;
-use core::mem;
-
 use crate::ArithmeticError;
+
+pub(crate) mod sqrt;
+
+use sqrt::Sqrt;
 
 pub trait Zero {
     const ZERO: Self;
@@ -324,43 +325,28 @@ pub trait RoundingDiv<Rhs = Self> {
 pub trait RoundingSqrt: Sized {
     type Error;
 
-    /// Checked rounded square root.
-    /// Returns `Err` for negative arguments.
+    /// Checked [rounding][RoundMode] square root.
+    /// Returns `Err` for negative argument.
     ///
     /// ```
-    /// use fixnum::{FixedPoint, typenum::U9, ops::{Zero, RoundingSqrt, RoundMode::*}};
+    /// use fixnum::{ArithmeticError, FixedPoint, typenum::U9};
+    /// use fixnum::ops::{Zero, RoundingSqrt, RoundMode::*};
     ///
     /// type Amount = FixedPoint<i64, U9>;
     ///
     /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
     /// let a: Amount = "81".parse()?;
+    /// let b: Amount = "2".parse()?;
+    /// let c: Amount = "-100".parse()?;
     /// assert_eq!(a.rsqrt(Floor)?, "9".parse()?);
+    /// assert_eq!(b.rsqrt(Floor)?, "1.414213562".parse()?);
+    /// assert_eq!(b.rsqrt(Ceil)?, "1.414213563".parse()?);
+    /// assert_eq!(c.rsqrt(Floor), Err(ArithmeticError::DomainViolation));
     /// # Ok(()) }
     /// ```
     ///
-    /// [FixedPoint]: ../struct.FixedPoint.html
     /// [RoundMode]: ./enum.RoundMode.html
     fn rsqrt(self, mode: RoundMode) -> Result<Self, Self::Error>;
-}
-
-pub trait Sqrt: Sized {
-    type Error;
-
-    /// Checked integer square root.
-    /// Returns `Err` for negative arguments.
-    ///
-    /// ```
-    /// use fixnum::ops::Sqrt;
-    ///
-    /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
-    /// assert_eq!(81.sqrt()?, 9);
-    /// assert_eq!(80.sqrt()?, 8);
-    /// # Ok(()) }
-    /// ```
-    ///
-    /// [FixedPoint]: ../struct.FixedPoint.html
-    /// [RoundMode]: ./enum.RoundMode.html
-    fn sqrt(self) -> Result<Self, Self::Error>;
 }
 
 // Impls for primitives.
@@ -452,85 +438,22 @@ macro_rules! impl_for_ints {
                 Ok(result)
             }
         }
+
+        impl RoundingSqrt for $int {
+            type Error = ArithmeticError;
+
+            #[inline]
+            fn rsqrt(self, mode: RoundMode) -> Result<Self, Self::Error> {
+                let lo = self.sqrt()?;
+                Ok(match mode {
+                    RoundMode::Floor => lo,
+                    RoundMode::Ceil => if lo * lo == self { lo } else {
+                        lo + <$int>::ONE
+                    },
+                })
+            }
+        }
     };
 }
 
 impl_for_ints!(i8, i16, i32, i64, i128); // TODO: unsigned?
-
-macro_rules! impl_sqrt {
-    ($( $int:ty ),+ $(,)?) => {
-        $( impl_sqrt!(@single $int); )*
-    };
-    (@single $int:ty) => {
-        impl Sqrt for $int {
-            type Error = ArithmeticError;
-
-            /// Implementation courtesy to `num` crate.
-            /// @see https://github.com/rust-num/num-integer/blob/4d166cbb754244760e28ea4ce826d54fafd3e629/src/roots.rs#L278
-            #[inline]
-            fn sqrt(self) -> Result<Self, Self::Error> {
-                #[inline]
-                const fn bits<T>() -> u32 {
-                    (mem::size_of::<T>() * 8) as _
-                }
-
-                #[inline]
-                fn fixpoint(mut x: $int, f: impl Fn($int) -> $int) -> $int {
-                    let mut xn = f(x);
-                    while x < xn {
-                        x = xn;
-                        xn = f(x);
-                    }
-                    while x > xn {
-                        x = xn;
-                        xn = f(x);
-                    }
-                    x
-                }
-
-                #[cfg(feature = "std")]
-                #[inline]
-                fn guess(x: $int) -> $int {
-                    (x as f64).sqrt() as $int
-                }
-
-                #[cfg(not(feature = "std"))]
-                #[inline]
-                fn guess(x: $int) -> $int {
-                    #[inline]
-                    fn log2_estimate(x: $int) -> u32 {
-                        debug_assert!(x > 0);
-                        bits::<$int>() - 1 - x.leading_zeros()
-                    }
-
-                    1 << ((log2_estimate(x) + 1) / 2)
-                }
-
-                #[allow(unused_comparisons)]
-                if self < 0 {
-                    return Err(ArithmeticError::DomainViolation);
-                }
-                if bits::<$int>() > 64 {
-                    // 128-bit division is slow, so do a recursive bitwise `sqrt` until it's small enough.
-                    let result = match u64::try_from(self) {
-                        Ok(x) => x.sqrt()? as _,
-                        Err(_) => {
-                            let lo = (self >> 2u32).sqrt()? << 1;
-                            let hi = lo + 1;
-                            if hi * hi <= self { hi } else { lo }
-                        }
-                    };
-                    return Ok(result);
-                }
-                if self < 4 {
-                    return Ok((self > 0).into());
-                }
-                // https://en.wikipedia.org/wiki/Methods_of_computing_square_roots#Babylonian_method
-                let next = |x: $int| (self / x + x) >> 1;
-                Ok(fixpoint(guess(self), next))
-            }
-        }
-    }
-}
-
-impl_sqrt!(i8, u8, i16, u16, i32, u32, i64, u64, i128, u128);